requirements-optional-diarization.txt
pyannote.audio
pyannote.core
pyannote.database
pyannote.metrics
pyannote.pipeline
pytorch-metric-learning
torch-audiomentations
torchcodec
